focal.mean(): only do data type conversion once (#699)
* focal.mean(): free memory after each iteration

* do type conversion once at first pass

+++ xrspatial/focal.py
@@ -100,7 +100,7 @@ def _mean_cupy(data, excludes):
 
 
 def _mean(data, excludes):
-    agg = xr.DataArray(data.astype(float))
+    agg = xr.DataArray(data)
     mapper = ArrayTypeFunctionMapping(
         numpy_func=_mean_numpy,
         cupy_func=_mean_cupy,
@@ -201,7 +201,7 @@ def mean(agg, passes=1, excludes=[np.nan], name='mean'):
         Dimensions without coordinates: dim_0, dim_1
     """
 
-    out = agg.data
+    out = agg.data.astype(float)
     for i in range(passes):
         out = _mean(out, tuple(excludes))
